+++ src/Orc.NuGetExplorer/NuGet/Packaging/EmptyPackageDetails.cs
@@ -48,7 +48,7 @@ public class EmptyPackageDetails : IPackageDetails
 
     public DateTimeOffset? Published { get; }
 
-    public int? DownloadCount { get; }
+    public long? DownloadCount { get; }
 
     public string Dependencies { get; }
